Revert "Update llvm to green 6cf7fe4a (#47)"
This reverts commit a7f1c0d5fd01836609e6143187d0cdbc51ce54a2.

+++ lldb/source/Plugins/Platform/gdb-server/PlatformRemoteGDBServer.cpp
@@ -721,7 +721,7 @@ const UnixSignalsSP &PlatformRemoteGDBServer::GetRemoteUnixSignals() {
           return false;
 
         // Signal number and signal name are required.
-        uint64_t signo;
+        int signo;
         if (!dict->GetValueForKeyAsInteger("signo", signo))
           return false;
